# src/main/resources/openapi/room-api.yaml
openapi: 3.0.3
info:
  title: Suite Sync Room Management API
  description: API for managing rooms and checking room availability in the hotel management system
  version: 1.0.0
  contact:
    name: Suite Sync Team
    email: support@suitesync.com

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Rooms
    description: Room management and availability operations

paths:
  /api/v1/rooms:
    get:
      tags:
        - Rooms
      summary: Get all rooms
      description: Retrieve a list of rooms, optionally filtered by availability dates and number of guests
      operationId: getAllRooms
      security: []
      parameters:
        - name: checkInDate
          in: query
          description: Check-in date to filter available rooms
          required: false
          schema:
            type: string
            format: date
            example: 2023-12-20
        - name: checkOutDate
          in: query
          description: Check-out date to filter available rooms
          required: false
          schema:
            type: string
            format: date
            example: 2023-12-25
        - name: numberOfGuests
          in: query
          description: Number of guests to filter rooms by capacity
          required: false
          schema:
            type: integer
            minimum: 1
            example: 2
        - name: roomType
          in: query
          description: Filter rooms by type
          required: false
          schema:
            $ref: '#/components/schemas/RoomType'
        - name: minPrice
          in: query
          description: Minimum price per night
          required: false
          schema:
            type: number
            format: double
            minimum: 0
            example: 50.00
        - name: maxPrice
          in: query
          description: Maximum price per night
          required: false
          schema:
            type: number
            format: double
            minimum: 0
            example: 500.00
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomPageResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Rooms
      summary: Create a new room
      description: Create a new room in the system (admin/staff only)
      operationId: createRoom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Room with this number already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/rooms/{roomId}:
    get:
      tags:
        - Rooms
      summary: Get room by ID
      description: Retrieve detailed information about a specific room
      operationId: getRoomById
      security: []
      parameters:
        - name: roomId
          in: path
          description: ID of the room to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Rooms
      summary: Update room
      description: Update an existing room's information (admin/staff only)
      operationId: updateRoom
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          description: ID of the room to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: Room updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Rooms
      summary: Delete room
      description: Delete a room from the system (admin only)
      operationId: deleteRoom
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          description: ID of the room to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Room deleted successfully
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot delete room with active bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/rooms/{roomId}/availability:
    get:
      tags:
        - Rooms
      summary: Check room availability
      description: Check if a specific room is available for given dates
      operationId: checkRoomAvailability
      security: []
      parameters:
        - name: roomId
          in: path
          description: ID of the room to check
          required: true
          schema:
            type: integer
            format: int64
        - name: checkInDate
          in: query
          description: Check-in date
          required: true
          schema:
            type: string
            format: date
            example: 2023-12-20
        - name: checkOutDate
          in: query
          description: Check-out date
          required: true
          schema:
            type: string
            format: date
            example: 2023-12-25
      responses:
        '200':
          description: Availability check completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
        '400':
          description: Invalid date parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    RoomType:
      type: string
      enum:
        - SINGLE
        - DOUBLE
        - SUITE
        - DELUXE
        - PENTHOUSE
      description: Type of room

    RoomStatus:
      type: string
      enum:
        - AVAILABLE
        - OCCUPIED
        - MAINTENANCE
        - OUT_OF_SERVICE
      description: Current status of the room

    CreateRoomRequest:
      type: object
      required:
        - roomNumber
        - roomType
        - maxOccupancy
        - pricePerNight
        - size
      properties:
        roomNumber:
          type: string
          minLength: 1
          maxLength: 10
          description: Unique room number
          example: "101"
        roomType:
          $ref: '#/components/schemas/RoomType'
        maxOccupancy:
          type: integer
          minimum: 1
          maximum: 10
          description: Maximum number of guests the room can accommodate
          example: 2
        pricePerNight:
          type: number
          format: double
          minimum: 0
          description: Price per night in USD
          example: 150.00
        size:
          type: number
          format: double
          minimum: 0
          description: Room size in square meters
          example: 25.5
        floor:
          type: integer
          minimum: 0
          description: Floor number where the room is located
          example: 1
        description:
          type: string
          maxLength: 1000
          description: Detailed description of the room
          example: A comfortable room with a queen-sized bed and city view
        amenities:
          type: array
          items:
            type: string
          description: List of amenities available in the room
          example: ["WiFi", "TV", "Air Conditioning", "Mini Bar", "Safe"]
        images:
          type: array
          items:
            type: string
            format: uri
          description: URLs of room images
          example: ["https://example.com/room101-1.jpg", "https://example.com/room101-2.jpg"]

    UpdateRoomRequest:
      type: object
      properties:
        roomType:
          $ref: '#/components/schemas/RoomType'
        maxOccupancy:
          type: integer
          minimum: 1
          maximum: 10
          description: Maximum number of guests the room can accommodate
          example: 2
        pricePerNight:
          type: number
          format: double
          minimum: 0
          description: Price per night in USD
          example: 150.00
        size:
          type: number
          format: double
          minimum: 0
          description: Room size in square meters
          example: 25.5
        floor:
          type: integer
          minimum: 0
          description: Floor number where the room is located
          example: 1
        description:
          type: string
          maxLength: 1000
          description: Detailed description of the room
          example: A comfortable room with a queen-sized bed and city view
        status:
          $ref: '#/components/schemas/RoomStatus'
        amenities:
          type: array
          items:
            type: string
          description: List of amenities available in the room
          example: ["WiFi", "TV", "Air Conditioning", "Mini Bar", "Safe"]
        images:
          type: array
          items:
            type: string
            format: uri
          description: URLs of room images
          example: ["https://example.com/room101-1.jpg", "https://example.com/room101-2.jpg"]

    RoomResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the room
          example: 1
        roomNumber:
          type: string
          description: Room number
          example: "101"
        roomType:
          $ref: '#/components/schemas/RoomType'
        maxOccupancy:
          type: integer
          description: Maximum number of guests the room can accommodate
          example: 2
        pricePerNight:
          type: number
          format: double
          description: Price per night in USD
          example: 150.00
        size:
          type: number
          format: double
          description: Room size in square meters
          example: 25.5
        floor:
          type: integer
          description: Floor number where the room is located
          example: 1
        description:
          type: string
          description: Detailed description of the room
          example: A comfortable room with a queen-sized bed and city view
        status:
          $ref: '#/components/schemas/RoomStatus'
        amenities:
          type: array
          items:
            type: string
          description: List of amenities available in the room
          example: ["WiFi", "TV", "Air Conditioning", "Mini Bar", "Safe"]
        images:
          type: array
          items:
            type: string
            format: uri
          description: URLs of room images
          example: ["https://example.com/room101-1.jpg", "https://example.com/room101-2.jpg"]
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the room was created
          example: 2023-01-15T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the room was last updated
          example: 2023-01-15T10:30:00Z

    RoomPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RoomResponse'
        page:
          type: integer
          description: Current page number
          example: 0
        size:
          type: integer
          description: Number of items per page
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Total number of rooms
          example: 50
        totalPages:
          type: integer
          description: Total number of pages
          example: 3

    AvailabilityResponse:
      type: object
      properties:
        available:
          type: boolean
          description: Whether the room is available for the specified dates
          example: true
        roomId:
          type: integer
          format: int64
          description: ID of the room
          example: 1
        checkInDate:
          type: string
          format: date
          description: Requested check-in date
          example: 2023-12-20
        checkOutDate:
          type: string
          format: date
          description: Requested check-out date
          example: 2023-12-25
        totalPrice:
          type: number
          format: double
          description: Total price for the stay
          example: 750.00
        pricePerNight:
          type: number
          format: double
          description: Price per night
          example: 150.00
        numberOfNights:
          type: integer
          description: Number of nights for the stay
          example: 5

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error
          example: 2023-01-15T10:30:00Z
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: Bad Request
        message:
          type: string
          description: Error message
          example: Invalid input data
        path:
          type: string
          description: Request path
          example: /api/v1/rooms

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT